feat: App 컴포넌트의 로또 구매 로직 service 레이어로 분리 및 적용
