+++ src/components/App.js
@@ -6,9 +6,6 @@ import WinningNumberForm from './lottoWinningNumber/WinningNumberForm';
 import RewardModalInner from './lottoRewardResult/RewardModalInner';
 
 import Modal from './utils/Modals';
-import { getRandomNumber } from '../utils';
-
-import { LOTTO } from '../constants/lotto';
 
 import { Main, MainWrapperDiv, WidthFullDiv } from './App.style';
 
@@ -52,20 +49,6 @@ class App extends Component {
     });
   };
 
-  createLottos = price => {
-    const lottoCount = Math.floor(price / LOTTO.UNIT);
-
-    const newLottos = Array.from({ length: lottoCount }, () => {
-      const lotto = new Set();
-      while (lotto.size < LOTTO.BUNDLE_SIZE) {
-        lotto.add(getRandomNumber(LOTTO.START_NUM, LOTTO.END_NUM));
-      }
-      return [...lotto].sort((a, b) => a - b);
-    });
-
-    this.setLottos(newLottos);
-  };
-
   render() {
     const { lottos, isModalOpened } = this.state;
     const isPurchased = lottos.length > 0;
@@ -76,7 +59,7 @@ class App extends Component {
         <MainWrapperDiv>
           <WidthFullDiv>
             <LottoPurchaseForm
-              createLottos={this.createLottos}
+              setLottos={this.setLottos}
               ref={this.purchaseFormRef}
             />
